+++ variables.tf
@@ -11,18 +11,6 @@ variable "log_analytics_workspace" {
   nullable    = false
 }
 
-variable "webhook_name" {
-  type        = string
-  description = "Name of the alert webhook."
-  nullable    = false
-}
-
-variable "webhook_service_uri" {
-  type        = string
-  description = "Link to the webhook receiver URL."
-  nullable    = false
-}
-
 variable "additional_queries" {
   type = map(object({
     query_path  = string
@@ -34,10 +22,11 @@ variable "additional_queries" {
   nullable    = false
 }
 
-variable "event_pipeline_key" {
+variable "secret" {
   type        = string
-  description = "Function key provided by monitoring team."
+  description = "Value that will replace the placeholder `{{secret}}` in `event_pipeline_config.service_uri`."
   sensitive   = true
+  default     = ""
 }
 
 variable "automation_account" {
@@ -49,4 +38,35 @@ variable "automation_account" {
   })
   description = "Automation account where the resource graph script will be deployed."
   nullable    = false
+}
+
+variable "event_pipeline_config" {
+  type = object({
+    enabled     = bool
+    name        = optional(string, "QBY EventPipeline")
+    service_uri = optional(string)
+  })
+
+  description = <<-DOC
+  ```
+  {
+    enabled       = Enable the action group if you want to send data to a monitoring service.
+    name          = Name of the alert webhook.
+    service_uri   = Link to the webhook receiver URL. Must contain the placeholder \"{{secret}}\". This placeholder will be replaced by the secret value from var.secret. This is used to add authentication to the webhook URL as a query parameter.
+}
+  ```
+  DOC
+  default = {
+    enabled = false
+  }
+
+  validation {
+    condition     = var.event_pipeline_config.enabled ? var.event_pipeline_config.service_uri != null : true
+    error_message = "If the config is enabled the service_uri must be set."
+  }
+  
+  validation {
+    condition     = var.event_pipeline_config.enabled ? strcontains(var.event_pipeline_config.service_uri, "{{secret}}") : true
+    error_message = "If the config is enabled the service_uri must contain the {{secret}} placeholder."
+  }
 }