+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this module will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - 2024-01-02
+
+### Added
+ - New example for new feature
+ 
+### Changed
+ - Put all action group configuration into one variable
+ - Made the webhook configuration optional
+
+### Removed
+
+### Fixed
+ - Removed unused parameter for secret
+
 ## [2.0.3] - 2023-12-07
 
 ### Added